+++ packages/storefront/src/lib/composables/use-shared-data.ts
@@ -28,6 +28,4 @@ window.$storefront.data['${field}'] = ${JSON.stringify(value)};`,
 
 export default useSharedData;
 
-export {
-  useSharedData,
-};
+export { useSharedData };